refactor(api): move user login model
